Move to electron-builer rather than electron-forge
electron forge was harder to debug, and seemed to not support bundling
of os specific files.

As part of this, we need to link the build location of the python app
into a place the electron-builder config can understand.

We do this by setting up an appropriate environment variable in the just
file.

Not very clean, but it works

Also, added a basic icon