Secure the network traffic a bit more
Firstly, we use a random free port.

Secondly, we protect the server by a) binding to 127.0.0.1 only and b)
generating a random token on start up, and configuring the server to
require it, and set it as a cookie.

Longer term, we'd add TLS to the connection, but that's a bit trickier to
pull off. In theory, on start up, we can generate a self signed keypair,
dyanmically add the public key to the appstrust store, and get the
python app to run with that keypair. But that's a problem for another
day.