Move to django web service
This is a bit trickier. In a lot of places, django relies on __file__
being set.  However, we'd configured pyoxidizer to build in all the code
into the executable, which means that __file__ is not set.

This change switches pyoxidizer to ship the code on the filesystem
instead. Because we then package this up with electron, that works ok.

It's still not 100% finished yet - we need to sort out db location and
migrations for example - but it seems to work. And it gives us a regular
django project to work on.