Support relative paths in ACRO metadata
This adds support for ACRO JSON to have file paths relative to the
directory the JSON is in, to support viewing files on different machines
from the one that generated the outputs.

ACRO doesn't do this yet, but to make out test outputs portable and thus
useful for testing, we convert them to relative paths manually. We also
additionally copy custom outputs into the outputs/ dir, and make them a
relative path.

In the process, slightly refactored the contents url to pass the output
name rather than the output path, as that was cleaner and probably
should have been done in the first place.