Handle missing files when building zipfile
Instead of failing, record any missing files in a missing-files.txt
file.

In the process of writing tests for this, I uncovered a nasty bug where
I was POSTing but reading from the GET data. This bug was hidden due to
us defaulting on the path if not present in the GET request, but too
broadly applied.  This change limits that behaviour to just the index
view. We regret the error.

Fixes #85