POST explicit list of files to review endpoint
For now, it just posts all of them. Future changes will allow selection.

The backend validates and then generates a zip with just files the
requested.

The JS works in the simplest way possible, inserting the files when the
form is submitted. This plays very nicely with CSRF and
Content-Disposition downloads.