Parse and transform ACRO JSON into something more useful.
This make the JSON easier to consume in the UI code, and also provides a
single tested point to parse data or handle changes in format.

The only b/w incompat was renaming the filed containing the file path
from `output` to `path`, and the name output seems a little overused
already.

Fixes #99