Add a Cypress e2e test
This adds a "happy path" test, which successfully approves all files, releases
them and checks that a zipfile is returned.

It can be tested using npm run cypress:open and then clicking through to the
spec file. This needs the server to be running in the background.

There's a headless mode that can be used with npm run cypress:run too.