Invoke a modal when clicking the Download and release button
We're planning to do a lot more once all outputs are reviewed and the
current thinking is to put that into a modal.  At least for the upcoming
"global" comment.

This adds the modal, as a component to keep the implementation details
of that abstracted from the main template, invoked when the "Download
and release" button is clicked.  I've split the functionality for
enabling that button and submitting the form into two files now that
they're conceptually different parts of the UI.