Break the template down into a base and child template
We're going to be adding another template for the summary page but want
to make this change now so as to avoid holding up concurrent work on the
main viewer page.

I've pulled the header out into a component because we'll want to
maintain that for the second page but not have the Release and download
button so optional children was the ideal solution.