Add support for multiple files per output
This was a fairly deep change across both js/python code, undoing a lot
of assumptions.

But it now shows multiple files, and includes all file for an output in
the downloaded zip.

There's a caveat that we don't try and render the `outcome` overlay if
there is >1 file, as we currently don't know which file it applies to.

Finally managed to kill the old content_urls stuff too.

The filename display needs some styling, but its functional.