reset button styles when an output's comment is cleared
This reworks how we enable&disable buttons so the check for there being
content in the comment element moves into the keyup handler.  This
allows us to shortcut the logic when there *is* content, enabling both
buttons, and exiting the function.  Then when there is no content we
reset the state of the buttons back to their starting state, and disable
them as necessary based on the state.