Add link to open file in native application.
This adds download links to each file. With a regular browser, this will
download the file as usual, and can be opened.

In the electron app, we intercept the download, and then open it using
the system's native application for that file type. Because the user is
opening the downloaded copy, they cannot edit the original file via this
route, by design.