Build executable for linux
The AppImage didn't work for me, as it required the Python executable to be run
separately. However, the .deb seems to install and run as expected.

Building a .deb has some slightly different requirements than building a
Windows MSI, including requiring an email address for the author and a product
name that is all lowercase.