Update to ACRO 0.4.0 and handle new structure
The JSON has a top new level dict containing `version`, and the old data
in `results`.

For each output dict, the `output` key has been changed to be called
`files`, and is now a list of dicts not strings. The dicts contain the
file name, and the new 'sdc' per-cell structure (which we're not
currently using)

Additonaly, the tool now outptus a config.json file with the risk
profile data, and a checksums directory.

This change solidifies the loading of this data, exposing the `version`
and `config` as new top level fields to the UI, but preserving `outputs`
as the same dict of outputs.

Various bits of parsing, iterating and test data needed updating as a
results. The tranform function was no longer doing anything useful, so I
got rid. The Outputs class has subsumed its functionality.