Show a blank page when someone cancels the file open dialog at startup
This was leaving the app on the loading splash screen previously, but
showing an error page here isn't quite correct either since this is not
a true error because the user can always use the File menu to open the
dialog again.

We've opted for a blank page for now so that we can quickly avoid using
the error page.