Add support for directories of files without ACRO metdata
This means changing to use a directory picker, rather than a file
picker. This means a bit of guess work when there are acro metadata
files in the directory. The upshot of this is:

a) we now launch the app in a special /load url to do this guess work
b) we no longer support a directory with multiple acro metadata files in